% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipar-functions.R
\name{default.knots}
\alias{default.knots}
\alias{newtonr}
\alias{eta_f}
\alias{EM_f}
\alias{lambda1f}
\alias{lambda3f}
\alias{lambdanuf}
\title{Internal functions for smoothing}
\usage{
default.knots(x, num.knots)

newtonr(lambda2, eta, W2, R, tol = 10^(-2), maxiter = 25)

eta_f(lambda2, eta, W2, R, tol = 10^(-2), maxiter = 2)

EM_f(
  beta.hat,
  sigma,
  lambda1,
  lambda3,
  lambdanu,
  D,
  tol = 10^(-2),
  maxiter = 25
)

lambda1f(lambda, lambda3, lambdanu, sigma, beta, maxiter = 3)

lambda3f(nlambda = 25, lambda1, lambdanu, sigma, beta)

lambdanuf(nlambda = 25, sigma, lambda1)
}
\arguments{
\item{x, num.knots}{internal arguments}

\item{lambda2, eta, W2, R, tol, maxiter}{internal arguments}

\item{beta.hat, sigma, lambda1, lambda3, lambdanu, D, tol, maxiter}{internal arguments}

\item{lambda, lambda3, lambdanu, sigma, beta, maxiter}{internal arguments}

\item{nlambda, lambda1, lambdanu, sigma, beta}{internal arguments}

\item{nlambda, sigma, lambda1}{internal arguments}
}
\description{
These are not supposed to be used by others.
}
