% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollArima.R
\name{rollArima}
\alias{rollArima}
\title{Rolling Arima model fitting}
\usage{
rollArima(arima.model, ynew, horizon = 1)
}
\arguments{
\item{arima.model}{a time series or time series model for which forecasts are required.}

\item{ynew}{vector of response variables.}

\item{horizon}{number of periods for forecasting.}
}
\value{
A vector of forecasts from a rolling Arima model.
}
\description{
Rolling Arima model fitting
}
